Refactor gateway transformer and execCtx, add DownloadTable method commit_hash:0365ddbdb9634ccf726c5114b7dc2128deab7ab8
