+++ yt/yql/providers/yt/gateway/fmr/ya.make
@@ -18,6 +18,7 @@ PEERDIR(
     yt/yql/providers/yt/fmr/file/metadata/interface
     yt/yql/providers/yt/fmr/file/upload/interface
     yt/yql/providers/yt/lib/config_clusters
+    yt/yql/providers/yt/lib/url_mapper
     yt/yql/providers/yt/lib/schema
     yt/yql/providers/yt/provider
 )
