+++ yt/yql/providers/yt/gateway/fmr/yql_yt_fmr.cpp
@@ -13,6 +13,7 @@
 #include <yt/yql/providers/yt/fmr/process/yql_yt_job_fmr.h>
 #include <yt/yql/providers/yt/lib/lambda_builder/lambda_builder.h>
 #include <yt/yql/providers/yt/lib/schema/schema.h>
+#include <yt/yql/providers/yt/lib/url_mapper/yql_yt_url_mapper.h>
 #include <yt/yql/providers/yt/provider/yql_yt_helpers.h>
 
 #include <yql/essentials/core/yql_type_helpers.h>
@@ -57,6 +58,7 @@ public:
     {
         if (fmrServices->Config) {
             Clusters_ = MakeIntrusive<TConfigClusters>(*FmrServices_->Config);
+            UrlMapper_ = std::make_shared<TYtUrlMapper>(*FmrServices_->Config);
         }
 
         auto getOperationStatusesFunc = [this] {
@@ -628,7 +630,7 @@ public:
         Coordinator_->OpenSession(openRequest).GetValueSync();
 
         with_lock(Mutex_) {
-            Sessions_[sessionId] = MakeIntrusive<TFmrSession>(sessionId, options.UserName(), options.RandomProvider());
+            Sessions_[sessionId] = MakeIntrusive<TFmrSession>(sessionId, options.UserName(), options.RandomProvider(), options.TimeProvider(), options.OperationOptions(), options.ProgressWriter());
         }
         YQL_CLOG(INFO, FastMapReduce) << "Registered session " << sessionId << " with coordinator";
 
@@ -1075,7 +1077,7 @@ private:
     {
         TFmrSession::TPtr session = Sessions_[sessionId];
 
-        auto ctx = MakeIntrusive<TExecContextSimple<TOptions>>(FmrServices_, Clusters_, MkqlCompiler_, std::move(options), cluster, session);
+        auto ctx = MakeIntrusive<TExecContextSimple<TOptions>>(TIntrusivePtr<TFmrYtGateway>(this), FmrServices_, Clusters_, MkqlCompiler_, std::move(options), UrlMapper_, cluster, session);
         return ctx;
     }
 
@@ -1114,6 +1116,7 @@ private:
     TConfigClusters::TPtr Clusters_;
     TIntrusivePtr<NCommon::TMkqlCommonCallableCompiler> MkqlCompiler_;
     IYtJobService::TPtr YtJobService_;
+    std::shared_ptr<TYtUrlMapper> UrlMapper_;
 };
 
 } // namespace
