+++ yt/yql/providers/yt/lib/temp_files/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    temp_files.cpp
+)
+
+PEERDIR(
+    yql/essentials/utils
+)
+
+YQL_LAST_ABI_VERSION()
+
+END()
