+++ yt/yql/providers/yt/provider/ya.make
@@ -132,6 +132,7 @@ PEERDIR(
     yt/yql/providers/yt/lib/row_spec
     yt/yql/providers/yt/lib/schema
     yt/yql/providers/yt/lib/skiff
+    yt/yql/providers/yt/lib/temp_files
     yt/yql/providers/yt/lib/yson_helpers
     yt/yql/providers/yt/opt
     yt/yql/providers/yt/gateway/qplayer
