+++ yt/yql/providers/yt/provider/yql_yt_forwarding_gateway.cpp
@@ -153,4 +153,8 @@ NThreading::TFuture<IYtGateway::TDumpResult> TYtForwardingGatewayBase::Dump(TDum
     return Slave_->Dump(std::move(options));
 }
 
+NThreading::TFuture<IYtGateway::TDownloadTableResult> TYtForwardingGatewayBase::DownloadTable(TDownloadTableOptions&& options) {
+    return Slave_->DownloadTable(std::move(options));
+}
+
 } // namspace NYql
