+++ yt/yql/providers/yt/provider/yql_yt_forwarding_gateway.h
@@ -79,6 +79,8 @@ public:
 
     NThreading::TFuture<TDumpResult> Dump(TDumpOptions&& options) override;
 
+    NThreading::TFuture<TDownloadTableResult> DownloadTable(TDownloadTableOptions&& options) override;
+
 protected:
     IYtGateway::TPtr Slave_;
 };
